Reorganize all codes to /src folder
