--- Status.java
@@ -1,4 +1,0 @@
-
-public enum Status {
-	OnGoing, Success, Error
-}
