--- TellerClient.java
@@ -1,104 +1,0 @@
-
-public class TellerClient {
-	private String port;
-	private String ip;
-	private Teller teller;
-	private Teller[] tellers;
-	
-	// Getters & Setters
-	public String getPort() {
-		return this.port;
-	}
-	public void setPort(String newPort) {
-		port = newPort;
-	}
-	
-	public Teller getTeller() {
-		return this.teller;
-	}
-	public Teller[] getTellers() {
-		return this.tellers;
-	}
-	public void setTeller(Teller newTeller) {
-		teller = newTeller;
-	}
-	
-	public String getIp() {
-		return this.ip;
-	}
-	public void serIp(String newIp) {
-		ip = newIp;
-	}
-	
-	// Client operation functions
-	public void connectClient() {
-		
-	}
-	public void moveClient() {
-		
-	}
-	public void start() {
-		
-	}
-	public void signIn() {
-		
-	}
-	public void signOut() {
-		
-	}
-	// Teller functions
-	public void addTeller() {
-		
-	}
-	public void deleteTeller() {
-		
-	}
-	public void createUser() {
-		
-	}
-	public void openLogs() {
-		
-	}
-	public void loginUserAccount() {
-		
-	}
-	// !
-	public void checkTeller() {
-		
-	}
-	
-	public void createAccount() {
-		
-	}
-	public void deleteAccount() {
-		
-	}
-	public void addUser() {
-		
-	}
-	public void deleteUser() {
-		
-	}
-	public void forgetPassword() {
-		
-	}
-	public void changePin() {
-		
-	}
-	public void transferAdmin() {
-		
-	}
-	public void accessATM() {
-		
-	}
-	
-	// Driver
-	public static void main(String[] args) {
-		// create new tellerclient
-		TellerClient client = new TellerClient();
-		 
-		
-	}
-}
-
-
