+++ src/test/AllTests.java
@@ -1,0 +1,10 @@
+package test;
+
+import org.junit.platform.suite.api.SelectClasses;
+import org.junit.platform.suite.api.Suite;
+
+@Suite
+@SelectClasses({}) // add test class here
+public class AllTests {
+
+}
